+++ data/statistics_around_the_clock.go
@@ -431,23 +431,18 @@ func ReCalculateAroundTheClockStatistics() (map[int]map[int]*models.StatisticsAr
 
 	s := make(map[int]map[int]*models.StatisticsAroundThe)
 	for _, leg := range legs {
-		if leg.ID == 20546 {
-			stats, err := CalculateAroundTheClockStatistics(leg.ID)
-			if err != nil {
-				return nil, err
-			}
-			for playerID, stat := range stats {
-				log.Printf(`UPDATE statistics_around_the SET darts_thrown = %d, score = %d, longest_streak = %d, total_hit_rate = %f, hit_rate_1 = %f,
-					hit_rate_2 = %f, hit_rate_3 = %f, hit_rate_4 = %f, hit_rate_5 = %f, hit_rate_6 = %f, hit_rate_7 = %f, hit_rate_8 = %f, hit_rate_9 = %f,
-					hit_rate_10 = %f, hit_rate_11 = %f, hit_rate_12 = %f, hit_rate_13 = %f, hit_rate_14 = %f, hit_rate_15 = %f, hit_rate_16 = %f, hit_rate_17 = %f,
-					hit_rate_18 = %f, hit_rate_19 = %f, hit_rate_20 = %f, hit_rate_bull = %f WHERE leg_id = %d AND player_id = %d;`,
-					stat.DartsThrown, stat.Score, stat.LongestStreak.Int64, stat.TotalHitRate, stat.Hitrates[1], stat.Hitrates[2], stat.Hitrates[3], stat.Hitrates[4], stat.Hitrates[5],
-					stat.Hitrates[6], stat.Hitrates[7], stat.Hitrates[8], stat.Hitrates[9], stat.Hitrates[10], stat.Hitrates[11], stat.Hitrates[12], stat.Hitrates[13],
-					stat.Hitrates[14], stat.Hitrates[15], stat.Hitrates[16], stat.Hitrates[17], stat.Hitrates[18], stat.Hitrates[19], stat.Hitrates[20], stat.Hitrates[25],
-					leg.ID, playerID)
-			}
-			s[leg.ID] = stats
+		stats, err := CalculateAroundTheClockStatistics(leg.ID)
+		if err != nil {
+			return nil, err
 		}
+		for playerID, stat := range stats {
+			log.Printf(`UPDATE statistics_around_the SET darts_thrown = %d, score = %d, longest_streak = %d, total_hit_rate = %f, hit_rate_1 = %f, hit_rate_2 = %f, hit_rate_3 = %f, hit_rate_4 = %f, hit_rate_5 = %f, hit_rate_6 = %f, hit_rate_7 = %f, hit_rate_8 = %f, hit_rate_9 = %f, hit_rate_10 = %f, hit_rate_11 = %f, hit_rate_12 = %f, hit_rate_13 = %f, hit_rate_14 = %f, hit_rate_15 = %f, hit_rate_16 = %f, hit_rate_17 = %f, hit_rate_18 = %f, hit_rate_19 = %f, hit_rate_20 = %f, hit_rate_bull = %f WHERE leg_id = %d AND player_id = %d;`,
+				stat.DartsThrown, stat.Score, stat.LongestStreak.Int64, stat.TotalHitRate, stat.Hitrates[1], stat.Hitrates[2], stat.Hitrates[3], stat.Hitrates[4], stat.Hitrates[5],
+				stat.Hitrates[6], stat.Hitrates[7], stat.Hitrates[8], stat.Hitrates[9], stat.Hitrates[10], stat.Hitrates[11], stat.Hitrates[12], stat.Hitrates[13],
+				stat.Hitrates[14], stat.Hitrates[15], stat.Hitrates[16], stat.Hitrates[17], stat.Hitrates[18], stat.Hitrates[19], stat.Hitrates[20], stat.Hitrates[25],
+				leg.ID, playerID)
+		}
+		s[leg.ID] = stats
 	}
 
 	return s, err