+++ CHANGELOG.md
@@ -6,10 +6,16 @@
 - Added match statistics for `9 Dart Shootout`
 - New convenience endpoints`/statistics/<type_id>/<from>/<to>`
 - More statistics to global statistics endpoint
+- New endpoints for getting player statistics and player history
+
+#### Changed
+- Writing of `9 Dart Shootout` score to database
 
 #### Fixed
 - Correctly calculate legs played and won for different statistics
 - Fixed calculating of matches and legs played and won for shootout
+- Synchronized `AddVisit` function to prevent multiple entries of same score
+- Graceful handing of matches with venue id 0
 
 ## [1.0.0] - 2020-05-03
 #### Feature